feat(aurora-tools): look up tags on cluster if none supplied in config
Adds functionality to allow using tagging functions in a more generic way. If the tags object passed
into the constructor is empty, the applyTags method will now look up the tags on the cluster and use
them as the base to apply to the newly created instance. Additionally, you can now set namespace to
'*' to disable filtering tagging operations based on a namespace.